Create pagination and basic caffItem endpoint

+++ CaffStore.Backend/CaffStore.Backend.Data/Entities/File.cs
@@ -14,13 +14,13 @@ namespace CaffStore.Backend.Dal.Entities
 
 		public DateTimeOffset CreatedAt { get; set; }
 
-		public int? CreatedById { get; set; }
+		public long? CreatedById { get; set; }
 
 		public User CreatedBy { get; set; }
 
 		public DateTimeOffset LastModifiedAt { get; set; }
 
-		public int? LastModifiedById { get; set; }
+		public long? LastModifiedById { get; set; }
 
 		public User LastModifiedBy { get; set; }
 	}
